Add PageView factory for dynamic and repeaters
