+++ src/Object/RepeaterPageView.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * @copyright 2016 Vladimir Jimenez
+ * @license   https://github.com/allejo/stakx/blob/master/LICENSE.md MIT
+ */
+
+namespace allejo\stakx\Object;
+
+class RepeaterPageView extends PageView
+{
+
+}
